feat: Include ESLint and extend rules
The ESLint binary is included in the `package.json`. It is no longer necessary to include it as a devDependency in all projects.

Some scripts provided by this module can be used to test the code of a project. See README.md for more details.

BREAKING CHANGES: Rules are added to match the behavior of older ESLint configurations used